+++ src/test/kotlin/com/sparta/tfbq/domain/question/service/AnswerServiceTest.kt
@@ -2,7 +2,7 @@ package com.sparta.tfbq.domain.question.service
 
 import com.sparta.tfbq.domain.answer.dto.request.AddAnswerRequest
 import com.sparta.tfbq.domain.answer.service.AnswerService
-import com.sparta.tfbq.domain.member.model.Member
+import com.sparta.tfbq.domain.member.Member
 import com.sparta.tfbq.domain.member.model.MemberRole
 import com.sparta.tfbq.domain.member.repository.MemberRepository
 import com.sparta.tfbq.domain.question.dto.request.AddQuestionRequest
